requirements.txt
flask
flask-cors
PyJWT
python-dotenv
numpy
pytest
pytest-cov
flake8
pyzmq==24.0.1
cryptography==39.0.0
flask-socketio==5.3.2
python-engineio==4.3.4
python-socketio==5.7.2
